refactor: shop-list directory -> shop
