--- packages/server/src/modules/shop-list/dto/create-shop-list.dto.ts
@@ -1,1 +1,0 @@
-export class CreateShopListDto {}
