--- packages/server/src/modules/shop-list/dto/update-shop-list.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateShopListDto } from './create-shop-list.dto';
-
-export class UpdateShopListDto extends PartialType(CreateShopListDto) {}
