--- packages/server/src/modules/shop-list/entities/shop-list.entity.ts
@@ -1,12 +1,0 @@
-export class ShopList {
-  id: number;
-  title: string;
-  type: string[];
-  rating: number;
-  ratingCount: number;
-  // location: string;
-  // distance: number;
-  tags: string[];
-  averagePrice: number;
-  // favorite: boolean;
-}
