+++ packages/server/src/modules/shop/dto/create-shop-list.dto.ts
@@ -1,0 +1,1 @@
+export class CreateShopListDto {}
