+++ packages/server/src/modules/shop/dto/update-shop-list.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateShopListDto } from './create-shop-list.dto';
+
+export class UpdateShopListDto extends PartialType(CreateShopListDto) {}
